fix(storybook): comments out call of `StateDecorator` as it's not used
After updating the storybook packages, use of "@sambego/storybook-state"
seems to be different. Current code based on StateDecorator breaks for
our new version of storybook. It's used in Pagination component only.
This use is also not proper as the state information is not being read
in the story component. So as a fix, this decorator call has been
removed.