fix(OathKeeper): asc sort on Amount column works
Cause: for ascending sort of the column, value to param sortBy was sent as +Amount.

Our backend needs this value as Amount (without +). This is different than JSON:API standard, but because of ease of backend side development, we chose to do this.

I've fixed it. Now just like Advocate list's sorts, OathTaker's sorts will not mention + for ascending sorts.

Resolves: #317